Fikset OLP tester, fjernet tester som ikke er relevanta
