+++ src/test/resources/olp/søknad-fra-frontend.json
@@ -2,7 +2,7 @@
   "soeknadId": "21707da8-a13b-4927-8776-c53399727b29",
   "soekerId": "11111111111",
   "journalposter": [
-    "9999"
+    "99998888"
   ],
   "mottattDato": "2020-10-12",
   "klokkeslett": "12:53",
